pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["Oskar Korczak <oskar.korczak@protonmail.com>"]
readme = "README.md"
packages = [
    { include = "shortener" },
]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django = "^4.2.2"
drf-yasg = "^1.21.5"
drf-yasg-stubs = "^0.1.4"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.1"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
pylint-django = "^2.5.3"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.1"}
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring"
]
load-plugins = "pylint_django"
django-settings-module = "urlshortener.settings"
ignore = "shortener/migrations"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
incremental = "false"
strict = "true"

[tool.django-stubs]
django_settings_module = "urlshortener.settings"